Rendering in ROI can be now specified with subpixel precision (issue 12 at googlecode).

+++ src/main/java/cz/cuni/lf1/lge/ThunderSTORM/rendering/ui/HistogramRenderingUI.java
@@ -33,7 +33,7 @@ public class HistogramRenderingUI extends AbstractRenderingUI {
         initPars();
     }
 
-    public HistogramRenderingUI(int sizeX, int sizeY) {
+    public HistogramRenderingUI(double sizeX, double sizeY) {
         super(sizeX, sizeY);
         initPars();
     }
